refactor(core): 🔨 split more verification methods
