+++ DomainDetective/DomainHealthCheck.Verification.Bimi.cs
@@ -1,0 +1,39 @@
+using DnsClientX;
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace DomainDetective {
+    public partial class DomainHealthCheck {
+        /// <summary>
+        /// Analyzes a raw BIMI record.
+        /// </summary>
+        /// <param name="bimiRecord">BIMI record text.</param>
+        /// <param name="cancellationToken">Token to cancel the operation.</param>
+        public async Task CheckBIMI(string bimiRecord, CancellationToken cancellationToken = default) {
+            await BimiAnalysis.AnalyzeBimiRecords(new List<DnsAnswer> {
+                new DnsAnswer {
+                    DataRaw = bimiRecord,
+                    Type = DnsRecordType.TXT
+                }
+            }, _logger, cancellationToken: cancellationToken);
+        }
+
+        /// <summary>
+        /// Queries and analyzes BIMI records for a domain.
+        /// </summary>
+        /// <param name="domainName">Domain to verify.</param>
+        /// <param name="cancellationToken">Token to cancel the operation.</param>
+        public async Task VerifyBIMI(string domainName, CancellationToken cancellationToken = default) {
+            if (string.IsNullOrWhiteSpace(domainName)) {
+                throw new ArgumentNullException(nameof(domainName));
+            }
+            domainName = NormalizeDomain(domainName);
+            UpdateIsPublicSuffix(domainName);
+            BimiAnalysis = new BimiAnalysis();
+            var bimi = await DnsConfiguration.QueryDNS($"default._bimi.{domainName}", DnsRecordType.TXT, cancellationToken: cancellationToken);
+            await BimiAnalysis.AnalyzeBimiRecords(bimi, _logger, cancellationToken: cancellationToken);
+        }
+    }
+}
